==> 01_CalculateWiderRegion.Rmd <==
---
title: "Extended Area Mounds"
author: "Adela Sobotkova"
date: "2024-06-05"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sf)
library(raster)
library(tidyverse)
library(mapview)
```


## Load Yambol mounds, rasters, and region
```{r} 
# Yambol mounds
Yam_mnds <- readRDS("data/Yam_dd_mnds.rds") # 1073 features
Yam_mnds %>% 
  group_by(Type) %>% 
  tally()

# rasters
#YT_elev <- raster("data/large/YT_elev32635.tif") # vertical
#Y_elev <- raster("data/large/Yelev32635.tif")

Y35 <- raster("data/large/Y_elev_ext.tif") # horizontal

# region admin
region <- read_sf("data/YamRegion.shp")
#Y_buf25 <- st_buffer(region, 25000)
Y_buf25 <- st_read("data/Y_region_buf25k.shp")
plot(Y35);plot(region, borders = "blue", add=T)
```

## Mounds from Wider area
```{r mm-all}
# map mounds ~ 10,000 points
m_all <- read_sf("data/MapMounds32635.shp")
m_all %>% group_by(MpSymbl) %>% tally()

# filter the most conservative symbol (sunburst) 4290 points
mm <- m_all %>% 
    filter(grepl("^Hairy brown", MpSymbl)) # deleting 'brown' gets all rayed symbols
```

### Northern map mounds Angel digitized in June 2024
this concerns mounds north of Yambol - mostly in Stara Planina -  that fall within 25 km buffer of Yambol border 
```{r}
mm_a <- read_sf("../YambolMoundAnalysis2023/data/MapMoundsA32635.geojson")
mm_a$Notes <- ""
mm <- mm %>% 
  rename(Id =identfr, createdBy=cretdBy, createdAt = FtrTmst,Latitude = Latitud, Longitude = Longitd, MapSymbol = MpSymbl, Notes = Note) %>% 
  dplyr::select(Id, createdBy, createdAt, Latitude, Longitude, MapSymbol, Notes) %>% 
  rbind(mm_a)

```

### Intersect map mounds with 25 km buffer of Yambol region
```{r}
# intersect filtered sunbursts with region buffer
mm_Yext <- st_intersection(mm, (Y_buf25$geometry %>% 
  st_difference(region$geometry)))

# are there duplicate IDs in the 1065 map mounds?
mm_Yext %>% 
  group_by(Id) %>% 
  count() %>% arrange(desc(n))

# save intermediate products
# st_write(mm_Yext, "output_data/Y25_mapmounds.shp", append = F)
# st_write(Y_buf25, "data/Y_region_buf25k.shp")

# Visualize additional mounds and their relief
plot(Y_buf25$geometry, border = "red");plot(mm_Yext$geometry, add = T);plot(region$geometry, add =T, col = "green")

plot(Y35, main = "Digitized map mounds within 25km buffer of Yambol Province");plot(region$geometry, add =T);plot(mm_Yext$geometry, add = T , cex = 0.1)

mapview(Y35)+ mapview(mm_Yext)+mapview(region)
hist(values(Y35))
```


## Prepare Map Mounds for Intervisibility

### Focus on mounds on edges
Calculate intervisibility for origin mounds among those 5 km to the border and the mapped mound and then add this result to the current intervisibility table. How many mounds outside Yambol can these mounds see?

```{r}
innerbuff <- st_buffer(region, -5000)
edge5km_mounds <- st_difference(Yam_mnds, innerbuff)

# SOUTHERN BORDER IS INTRACTABLE..so skipping mounds within 4500m of border 
edge5km_mounds<- edge5km_mounds %>% 
  filter(unclass(distBG) >4500) 

plot(region$geometry);plot(innerbuff, add = T)
plot(edge5km_mounds$geometry, add = T)

```

```{r inter-prep}
# origin point coordinates and height
Yam_mnds$HeightMax[is.na(Yam_mnds$HeightMax)] <- 0
origin <- cbind(st_coordinates(Yam_mnds), h = Yam_mnds$HeightMax)

# MUST HAVE NON-NA HEIGHT
edge5km_mounds$HeightMax[is.na(edge5km_mounds$HeightMax)] <- 0
origin <- cbind(st_coordinates(edge5km_mounds), h = edge5km_mounds$HeightMax)

# target points in map mounds
target <- cbind(st_coordinates(mm_Yext), h = 2) 
#plot(Y35); plot(mm_Yext$geometry, add = T);plot(edge5km_mounds$geometry, add = T, col = "red")

# target points for 8007, first inside then outside mounds
target2 <- rbind(cbind(st_coordinates(Yam_mnds), h = Yam_mnds$HeightMax), target) 


class(origin)
class(target)
# Start experiment
dim(origin) 
dim(target) 

rm(mm,m_all, innerbuff)
```



## Activate the functions needed
```{r}
source("scripts/VisibilityFunctions.R")
```

## Parallelize!
Parallelisation is much faster than loops but still takes ca 6-12 hours for the whole dataset of 1000 by 1000 mounds. Run if needed
```{r auto-foreach, eval = FALSE}
# install.packages("doParallel")
# install.packages("foreach")
library(foreach)
library(doParallel)
library(data.table)

detectCores() # 22 as some are multi-threaded
detectCores(logical = FALSE) # 16 real ones

cl <- makeCluster(4) # keep 4 cores so screensaver can run on W11; lowering to 10
registerDoParallel(cl)

on.exit({
  try({
   cat("Attempting to stop cluster\n")
   ?stopImplicitCluster()        # package: `doParallel`
   })
 })


# FOR TESTING PURPOSES REDUCE THE SIZE!
#origin <- origin[1:100,] # 10, 100 out of 341
origin <- origin[101:341,] # 10, 100 out of 341
origin <- cbind(Yam_mnds %>% 
  filter(TRAP == 8007) %>% 
  select(TRAP, HeightMax) %>% 
  st_coordinates(), h = 4.5)

#target <- target[1066:1206,]  # half of 1206 works for smaller datasets! Halve the current size
target <- target
target <- target2 # for 8007 calculation

# Start experiment
dim(origin) 
dim(target) 

# nesting both i and e with foreach 
ie_table <- foreach (i= 1:nrow(origin), # reduce the load instead of 1073 in nrow(origin)
                     .combine = 'rbind',
                     .packages = c("data.table", "raster")) %do% {  
  print(i)
  foreach( e= 1:nrow(target), .combine = 'rbind',.packages = c("data.table", "raster")) %dopar% {
  print(e)
  result <-  cansee(Y35, origin[i,1:2], target[e,1:2], h1 = origin[i,3], h2 = target[e,3])
  result_row <- c(i,e, result)
  }
}

# tell R that we don't need the processes anymore
stopCluster(cl)


dim(ie_table) # 1,151,329 calculations take ~12 hours (10x less than for loop) 
head(ie_table, n =20)
# saveRDS(ie_table, "output_data/Y8007_intervis.rds")
# saveRDS(ie_table, "output_data/Yam_241_ext_mnds_intervis.rds") # 10 * 1200 calculations worked, 5000 were NA
rm(ie_table)

```
### Combine the results from downscaled calculation
```{r}
# ie <- readRDS("output_data/Yam_100ext_mnds_intervis.rds") # first 100 mounds of origin and 1065 of target
ie_table # first 100 mounds of origin and remaining 141 mounds of target
res_one <- as_tibble(ie)
res_two <- as_tibble(ie_table)


tail(res_one)
tail(res_two)

# update and rename values in res_two$V1 to 100-341
res_two$V1 <-res_two$V1 + 100 

# rename values inside res_two$V2 and merge with res_one

res_two <- res_two %>% 
   rename(i = V1, e=V2, visibility = V3)

res_two$TRAPtarget <- mm_Yext$Id[res_two$e]
res_two$TRAPorigin <- edge5km_mounds$TRAP[res_two$i]

# for 8007 only 
res_two$TRAPtarget <- c(Yam_mnds$TRAP, mm_Yext$Id)
res_two$TRAPorigin <- 8007
saveRDS(res_two, "output_data/Y8007_intervis.rds")

# initial res_one processing, uncomment if you are recreating from scratch
# result_table <- res_one %>% 
#   rename(target = V2)  %>% 
#   rbind(res_two[,-2]) %>% 
#   rename(i = V1, e=target, visibility = V3)
  
# combine the two results: (1-100, and 101 - 341) res_one and res_two
result_table <- res_one %>%
  rbind(res_two)

result_table %>% 
  filter(visibility ==1)

# result_table$TRAPorigin <- edge5km_mounds$TRAP[result_table$i]
# result_table$TRAPtarget <- mm_Yext$Id[result_table$e]

result_table %>%  
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n)) %>% 
#  slice(1:10) %>% 
  left_join(edge5km_mounds, by = c("TRAPorigin"="TRAP")) %>%
  st_as_sf(crs = 32635) %>% 
  mapview(cex = "n" , zcol = "HeightMax") +mapview(Yam_mnds,cex = 0.1) 

saveRDS(result_table, "output_data/Yam_ext_mnds_intervis.rds") # all 341 border mounds intervisibility to 25km buffer 1206 external mounds

```
 A lovely  map of the true visibility of border mounds, when looking 'outside' the region! THe most prominent increase in intervisibility is in the very western tip of mounds and in southeast corner of the region. This way we are working with the true intervisibility. Previous winners still rule, because they have the largest viewshed being on SvIlijski vuzvyshennia, unobsured by Bakadzhitsite or Stara Planina (like RM mounds below the mnts), and attest to the fact that EBA people were selective and were encompassing the landscape!
 
```{r}
Y_out <- readRDS("output_data/Yam_ext_mnds_intervis.rds")
ext_vis_sorted <- Y_out %>% 
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n)) 
ext_vis_sorted 
```
New winners in the border region are `head(ext_vis_sorted$TRAPorigin, 10)` with the first five seeing additional 100 mounds

Adding up these out-of-region visibilities, the final visibility for Yambol mounds is an aggregate of the two measures.


## Final visibility review
```{r in_ext-vis}
Y_in <- readRDS("output_data/Yam_mnds_intervis.rds")
Y_vis <- Y_in %>% rbind(Y_out)
all_vis_sorted <- Y_vis %>% 
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n))

# the winners of outside-looking mounds
Y_out %>% 
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n)) %>% slice(1:10) %>% pull(TRAPorigin) -> outside10
Y_in %>% 
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n)) %>% slice(1:10) %>% pull(TRAPorigin) -> inside10
outside10[which(outside10%in%inside10)]

Y_vis %>% 
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n))%>% slice(1:10) %>% pull(TRAPorigin) -> all10

outside10[which(outside10%in%inside10)]
all10[all10%in%inside10]

all_vis_sorted %>% 
  mutate(log_vis = log10(n)) %>% 
  slice(1:100) %>% 
  left_join(Yam_mnds, by = c("TRAPorigin"="TRAP")) %>%
  st_as_sf(crs = 32635) %>% 
  mapview(cex = "log_vis" , zcol = "n") + mapview(Yam_mnds,cex = 0.1) + mapview(mm_Yext, cex = 0.1) + mapview(region, alpha = 1)
```

And here is a map with 100 mounds inside Yambol that can see the most mounds inside and within a buffer of 25km outside the region
