==> 04_Maps.Rmd <==
---
title: "Maps"
output: html_document
date: "2024-06-24"
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warning = FALSE)
library(sf)
library(raster)
library(tidyverse)
library(mapview)

```

### Maps

```{r}
path <- "C:/Users/au616760/OneDrive - Aarhus universitet/Documents/TRAP_Oxbow"
path <- "C:/Users/Adela/OneDrive/Desktop/TRAP_Oxbow"
# A bit of local infrastructure and topography
Y_towns <- read_sf(paste0(path, "/TopoData/Modern_settlements_Yambol_TRAP.shp"))
Y_rivers <- read_sf(paste0(path, "/TopoData/Rivers_Yambol_TRAP.shp"))
Tonzos <- read_sf(paste0(path, "/TopoData/Tonzos.shp")) %>% st_transform(32635)
Y_roads <- read_sf(paste0(path, "/TopoData/Modern_roads_Yambol_TRAP.shp"))
BG <- read_sf("data/Bulgaria_Border.shp")
region <- read_sf("data/YamRegion.shp")

Y_city <- Y_towns %>% 
  filter(Name_en == "Yambol")

Y_vis <- readRDS("output_data/Yam_mnds_intervis.rds")

# Yambol rasters
library(terra)
Y <- rast("data/large/Y_elev_ext.tif")
d <- st_make_grid(st_bbox(st_buffer(region, 5000)), n =1)
dem_Y5 <- crop(Y, d)
options(scipen=999)
```



## Map background

```{r}
tiff("figures/background.tiff", width = 8, height = 8, units = "in", res = 300)
#png("figures/background.png", width = 8, height = 8, units = "in", res = 300)
plot(dem_Y5, col = grey.colors(5, start = 1, end = 0)); 
plot(region$geometry, add =T); 
#plot(mm_Yext$geometry, cex = 0.4, add =T);
plot(BG$geometry, add = T, lwd = 4)
plot(Y_city$geometry, add = T, col = "darkgrey", border ="darkgrey")
plot(Tonzos$geometry, lwd = 2, col = "lightblue", add =T);
plot(Yam_mnds$geometry, cex= 0.4, add =T); 

# Add north arrow at coordinates (8, 8). See line 135
add_north_arrow(502000, 4642000, length = 2500, label = "N")
# Add scale bar at coordinates (2,502000, 4645000 1)
add_scale_bar(480000, 4640000)
text("Yambol", x = 460000, y = 4702000)

dev.off()
```

## MOST Intervisible 

```{r}
# group by viewpoint, join to spatial points and plot on a map
library(mapview)

Y_vis %>% 
  group_by(TRAPorigin) %>% 
  tally(visibility) %>% 
  arrange(desc(n)) %>% 
  slice(1:50) %>% 
  mutate(n_log = log10(n)) %>% 
  # join the summarized dataframe of 25 most visible mounds to mound spatial points
  left_join(Yam_mnds, by = c("TRAPorigin"="TRAP")) %>%
  # activate the geometry column with proper CRS
  st_as_sf(crs = 32635) %>% 
  # plot the 25 most visible mounds, varying their radius by number of visible features from this location and varying color by mound height
  mapview(cex = "n" , zcol = "n_log") + 
  # show other mounds in the background as small dots
  mapview(Yam_mnds,cex = 0.1) 
  # differentiate mounds visible from 9412 (the most visible single features)
  #mapview(Yam_mnds %>% dplyr::filter(TRAP%in%visiblefrom9412), cex = 1)

mapview(Yam_mnds[Yam_mnds$TRAP %in% visiblefrom9412,])
```

##  9412

```{r}
Y_vis %>% 
  filter(TRAPorigin == 9412, visibility == 1) %>% 
  arrange(desc(e)) %>% 
  distinct(TRAPtarget) %>% 
  pull() -> visiblefrom9412
```


### Visibility lines for 9412
```{r linestrings-9412}
# Test linestring creation from 9412 and its 409 visible mounds. The coords object needs to be a matrix, but to cbind one to many coordinates the component columns need to be dataframes.

mostvisiblemnd <- visiblefrom9412 # fill in model (veg10, veg20, none) for which we are calculating the lines of sight to the most visible mound

coords <- as.matrix(cbind(as.data.frame(st_coordinates(Yam_mnds %>% filter(TRAP == 9412))),
                as.data.frame(st_coordinates(Yam_mnds %>% filter(TRAP %in% mostvisiblemnd)))))
coords_out <- as.matrix(cbind(as.data.frame(st_coordinates(Yam_mnds %>% filter(TRAP == 9412))),
                as.data.frame(st_coordinates(mm_Yext %>% filter( Id %in% mostvisiblemnd)))))
coords <- rbind(coords, coords_out)
colnames(coords) <- c("X1", "Y1", "X2", "Y2")

lines_9412 <-  st_sfc(
     lapply(1:nrow(coords),
           function(i){
             st_linestring(matrix(coords[i,],ncol=2,byrow=TRUE))
           }))

st_crs(lines_9412) <- st_crs(Yam_mnds)


library(mapview)
mapview(lines_9412)+ 
  mapview(Yam_mnds %>% filter(TRAP %in% mostvisiblemnd))

# https://stackoverflow.com/questions/65498300/how-to-efficiently-create-linestrings-from-points
# https://stackoverflow.com/questions/58150279/plotting-lines-between-two-sf-point-features-in-r

```

### Figure 4
```{r}
# Start with map elements designed for 10in map offprint
# Function to add a north arrow
add_north_arrow <- function(x, y, length = 1, label = "N") {
  arrows(x, y, x, y + length, col = "black", lwd = 2)
  text(x, y + length + 1000, label, cex = 1, col = "black")
}

# Function to add a scale bar
add_scale_bar <- function(x, y, length = 25000, label = "25 km") {
  segments(x, y, x + length, y, col = "black", lwd = 2)
  segments(x, y, x, y + 1000, col = "black", lwd = 2)
  segments(x + length, y, x + length, y + 0.2, col = "black", lwd = 2)
  text(x + length / 2, y - 1000, label, cex = 1, col = "black")
}

```

```{r fig9412}
tiff("figures/04.tiff", width = 10, height = 10, units = "in", res = 300)
plot(dem_Y5, col = grey.colors(5, start = 1, end = 0)); plot(region$geometry, add =T); plot(Yam_mnds$geometry, cex= 0.4, add =T); plot(mm_Yext$geometry, cex = 0.4, add =T);
# plot(Y_rivers, col= "blue", add =T)
plot(BG$geometry, add = T, lwd = 4); plot(lines_9412, add =T)
# Add north arrow at coordinates (8, 8)
add_north_arrow(502000, 4642000, length = 2500, label = "N")
# Add scale bar at coordinates (2,502000, 4645000 1)
add_scale_bar(480000, 4640000)
dev.off()
```

### Visibility lines for 8007
```{r linestrings-8007}
# Test linestring creation from 9412 and its 409 visible mounds. The coords object needs to be a matrix, but to cbind one to many coordinates the component columns need to be dataframes.

Y8007 <- readRDS("output_data/Y8007_intervis.rds")
visiblefrom8007 <- Y8007 %>% 
  filter(visibility == 1) %>% 
  pull(TRAPtarget)
mostvisiblemnd <- visiblefrom8007 # fill in model (veg10, veg20, none) for which we are calculating the lines of sight to the most visible mound

coords <- as.matrix(cbind(as.data.frame(st_coordinates(Yam_mnds %>% filter(TRAP == 8007))),
                as.data.frame(st_coordinates(Yam_mnds %>% filter(TRAP %in% mostvisiblemnd)))))
coords_out <- as.matrix(cbind(as.data.frame(st_coordinates(Yam_mnds %>% filter(TRAP == 8007))),
                as.data.frame(st_coordinates(mm_Yext %>% filter( Id %in% mostvisiblemnd)))))
coords <- rbind(coords, coords_out)
colnames(coords) <- c("X1", "Y1", "X2", "Y2")


lines_8007 <-  st_sfc(
     lapply(1:nrow(coords),
           function(i){
             st_linestring(matrix(coords[i,],ncol=2,byrow=TRUE))
           }))

st_crs(lines_8007) <- st_crs(Yam_mnds)


library(mapview)
mapview(lines_8007)+ 
  mapview(Yam_mnds %>% filter(TRAP %in% mostvisiblemnd))

# https://stackoverflow.com/questions/65498300/how-to-efficiently-create-linestrings-from-points
# https://stackoverflow.com/questions/58150279/plotting-lines-between-two-sf-point-features-in-r

```

### Figure 5

```{r fig8007}
tiff("figures/05.tiff", width = 10, height = 10, units = "in", res = 300)
plot(dem_Y5, col = grey.colors(5, start = 1, end = 0)); 
plot(region$geometry, add =T); 
plot(Yam_mnds$geometry, cex= 0.4, add =T); 
plot(mm_Yext$geometry, cex = 0.4, add =T);
plot(BG$geometry, add = T, lwd = 4); 
plot(lines_8007, add =T)
# Add north arrow at coordinates (8, 8)
add_north_arrow(502000, 4642000, length = 2500, label = "N")
# Add scale bar at coordinates (2,502000, 4645000 1)
add_scale_bar(480000, 4640000)
dev.off()
```

## Intervisibility Distances
```{r}
# Check that you have 9412 coords
dim(coords)
coords
# Calculate distances
distances <- sqrt((coords[, "X2"] - coords[, "X1"])^2 + (coords[, "Y2"] - coords[, "Y1"])^2)
length(distances)

lines_9412_sf <- st_sf(geometry = lines_9412)


lines_9412_sf$length_m <- distances
lines_9412_sf$length_km <- distances/1000
hist(lines_9412_sf$length_km,
     xlab = "Distance (km)", 
     ylab = "Number of visible mounds", 
     main = "Mounds visible from 9412 (no = 491)")


lines_9412_sf$visibility_group <- cut(lines_9412_sf$length_km,
                                   breaks = c(0, 5, 10, 20, Inf),
                                   labels = c("1-5km", "5-10km", "10-20km","20+km"),
                                   right = FALSE)
head(lines_9412_sf$visibility_group)
head(lines_9412_sf$length_km)


line_types <- setNames(c("solid", "dashed", "dotted", "dotdash"), c("1-5km", "5-10km", "10-20km", "20+km"))


# Plot everything toether (run through row 264)
plot(dem_Y5, col = grey.colors(5, start = 1, end = 0)); 
plot(region$geometry, add =T); 
plot(Yam_mnds$geometry, cex= 0.4, add =T); 
plot(mm_Yext$geometry, cex = 0.4, add =T);
plot(BG$geometry, add = T, lwd = 4); 

# Plot lines by visibility group
# for (type in names(line_types)) {
#   lines_subset <- st_geometry(lines_9412_sf[lines_9412_sf$visibility_group == type, ])
#   plot(lines_subset, add = TRUE, lty = line_types[[type]])
# }
# Plot lines by visibility group
plot(st_geometry(lines_9412_sf[lines_9412_sf$visibility_group == "20+km", ]), add = TRUE, lty = "dotted")
plot(st_geometry(lines_9412_sf[lines_9412_sf$visibility_group == "10-20km", ]), add = TRUE, lty = "dashed")
plot(st_geometry(lines_9412_sf[lines_9412_sf$visibility_group == "5-10km", ]), add = TRUE, lty = "solid")
plot(lines_9412_sf %>% filter(visibility_group == "1-5km") %>% st_geometry(),  add = TRUE, lty = "solid")




# Add a legend
legend("topright", legend = names(line_types), lty = line_types, title = "Visibility Range")








for (type in names(line_types)) {
  lines_subset <- st_geometry(lines_9412_sf[lines_9412_sf$visibility_group == type, ])
  plot(lines_subset, add = TRUE, lty = line_types[[type]])
}



```

```{r}
# Define line styles for each visibility group
line_styles <- c("1-5km" = "solid", "5-10km" = "dashed", "10-20km" = "dotted", "20+km" = "dotdash")

# Plotting with ggplot2 for better control over aesthetics
ggplot() +
  geom_raster(data = dem_Y5, fill = grey.colors(5, start = 1, end = 0)) +  # Assuming dem_Y5 is also an sf object
  geom_sf(data = lines_9412_sf, aes(color = visibility_group, linetype = visibility_group)) +
  scale_linetype_manual(values = line_styles) +
  labs(title = "Visibility from Point 9412",
       color = "Visibility Range",
       linetype = "Visibility Range") +
  theme_minimal()
```
